requirements.txt
langchain
python-dotenv
openai
arxiv
pymupdf
langchain-text-splitters
langchain-openai
sentence-transformers
langchain-huggingface
faiss-cpu
chromadb
langchain-chromadb
langchain-google-genai
langchain_ollama
streamlit
langchain_groq